Complete and Upload to server

+++ database/migrations/2020_03_09_050045_create_products_table.php
@@ -21,9 +21,9 @@ class CreateProductsTable extends Migration
             $table->text('description');
             $table->string('slug');
             $table->integer('quantity')->default(1);
-            $table->integer('price');
+            $table->double('price');
             $table->tinyInteger('status')->default(0);
-            $table->integer('offer_price')->nullable();
+            $table->double('offer_price')->nullable();
             $table->integer('admin_id')->unsigned();
             $table->timestamps();
         });
